requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-multipart==0.0.6
pydantic==1.10.12
google-generativeai==0.5.0
python-dotenv==1.0.0